Move util/stats to generation/stats. Remove unused files left in pkg/util
