+++ generation/stats/combineprobs.go
